Set correct key and config for testing.
With the previous implementation, preedit text was transliterated to
get the query for prediction (half width ascii to full width) and this
was not the expected behavior.

By this change, we will be able to test the suggestion result of ascii
preedit text.

TODO(toshiyuki):
- Use new method(Composer::SetPreeditTextForTestOnly) for other test modules
or
- Modify the implementation of InsertCharacterPreedit() and remove
  SetPreeditTextForTestOnly() if we can confirm that that will not cause
  any problems.

REF_BUG=67485370
REF_CL=173622500
REF_TIME=2017-10-27T11:01:19+09:00
REF_TIME_RAW=1509069679 +0900