Fix warning: not handled in switch [-Wswitch]
When Mozc is built with Qt 5.5 or higher, you may see the following
warning as there is no handler for 'QtInfoMsg' message type.

  ../../gui/base/debug_util.cc:44:11: warning: enumeration value 'QtInfoMsg' not handled in switch [-Wswitch]
    switch (type) {
          ^
  1 warning generated.

This CL fixes the above warning by explicitly adding 'QtInfoMsg'
handler if the Qt version is 5.5 or higher.

Perhaps we may consider adding "default" case there so that we don't
see this warning again, but I decided not to do that (at least in this
CL) as this warning was actually useful (at least for me).

REF_BUG=
REF_CL=180312466
REF_TIME=2017-12-28T19:50:48-08:00
REF_TIME_RAW=1514519448 -0800